Jit64: Make TrampolineInfo smaller
Combined with the previous commit, this brings the TrampolineInfo struct
down to 48 bytes. This matters, because Jit64 has a big
std::unordered_map where it stores many megabytes of TrampolineInfo
entries.

The key saving comes from shrinking the len member from u32 to u16. It
should be safe to even turn it into a u8, but going that far brings no
additional savings due to how the padding works out.

+++ Source/Core/Core/PowerPC/Jit64Common/EmuCodeBlock.cpp
@@ -355,7 +355,7 @@ void EmuCodeBlock::SafeLoadToReg(X64Reg reg_value, const Gen::OpArg& opAddress, 
     {
       NOP(padding);
     }
-    info.len = static_cast<u32>(GetCodePtr() - info.start);
+    info.len = static_cast<u16>(GetCodePtr() - info.start);
 
     js.fastmemLoadStore = mov.address;
     return;
@@ -527,7 +527,7 @@ void EmuCodeBlock::SafeWriteRegToReg(OpArg reg_value, X64Reg reg_addr, int acces
     {
       NOP(padding);
     }
-    info.len = static_cast<u32>(GetCodePtr() - info.start);
+    info.len = static_cast<u16>(GetCodePtr() - info.start);
 
     js.fastmemLoadStore = mov.address;
